third_party/WebKit/LayoutTests/fast/js/Promise-static-every-expected.txt

Test Promise.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS result is undefined
PASS Promise.every() is fulfilled.
PASS result is undefined
PASS Promise.every(p1, p2, p3) is fulfilled.
PASS result.length is 3
PASS result[0] is "p1"
PASS result[1] is "p2"
PASS result[2] is "p3"
PASS Promise.every(p1, p6, p5) is rejected.
PASS result is "p6"
PASS Promise.every(p9) is fulfilled.
PASS result.length is 1
PASS result[0] is "p2"
PASS successfullyParsed is true

TEST COMPLETE
